requirements/base.txt
aniso8601
asgiref
attrs
black
cffi
click
cryptography
cycler
Django
django-environ
django-extensions
django-filter
django-object-actions
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
drf-jwt
drf-nested-routers
drf-spectacular
fonttools
graphene
graphene-django
graphql-core
graphql-relay
inflection
iso8601
jsonschema
kiwisolver
mypy-extensions
numpy
packaging
pandas
pathspec
Pillow
pip
platformdirs
promise
pycparser
PyJWT
pyparsing
pyrsistent
python-dateutil
pytz
PyYAML
Rx
setuptools
singledispatch
six
sqlparse
text-unidecode
tomli
typing-extensions
uritemplate
whitenoise
